rename to schema/openapi directory
